+++ app/src/main/java/com/hamdan/forzenbook/view/login/LoginActivity.kt
@@ -10,18 +10,20 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.hamdan.forzenbook.compose.core.theme.ForzenBookTheme
+import com.hamdan.forzenbook.createaccount.core.viewmodel.toCreateAccountUiState
 import com.hamdan.forzenbook.login.compose.MainLoginContent
+import com.hamdan.forzenbook.login.core.viewmodel.toLoginUiState
 import com.hamdan.forzenbook.view.LocalNavController
 import com.hamdan.forzenbook.view.NavigationDestinations
+import com.hamdan.forzenbook.viewmodels.CreateAccountViewModel
 import com.hamdan.forzenbook.viewmodels.LoginViewModel
-import com.hamdan.forzenbook.viewmodels.toCreateAccountUiState
-import com.hamdan.forzenbook.viewmodels.toLoginUiState
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class LoginActivity : ComponentActivity() {
 
     private val loginViewModel: LoginViewModel by viewModels()
+    private val createAccountViewModel: CreateAccountViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -36,7 +38,7 @@ class LoginActivity : ComponentActivity() {
                     ) {
                         composable(NavigationDestinations.LOGIN_PAGE) {
                             MainLoginContent(
-                                state = loginViewModel.loginState.value.toLoginUiState(),
+                                state = loginViewModel.state.value.toLoginUiState(),
                                 onInfoDismiss = {
                                     loginViewModel.loginDismissInfoClicked()
                                 },
@@ -59,12 +61,12 @@ class LoginActivity : ComponentActivity() {
                         }
                         composable(NavigationDestinations.CREATE_ACCOUNT) {
                             CreateAccountContent(
-                                state = loginViewModel.createAccountState.value.toCreateAccountUiState(),
+                                state = createAccountViewModel.state.value.toCreateAccountUiState(),
                                 onErrorDismiss = {
-                                    loginViewModel.createAccountDismissErrorClicked()
+                                    createAccountViewModel.createAccountDismissErrorClicked()
                                 },
                                 onTextChange = { firstName, lastName, birthDate, email, location ->
-                                    loginViewModel.updateCreateAccountTextAndErrors(
+                                    createAccountViewModel.updateCreateAccountTextAndErrors(
                                         firstName,
                                         lastName,
                                         birthDate,
@@ -72,11 +74,11 @@ class LoginActivity : ComponentActivity() {
                                         location,
                                     )
                                 },
-                                onDateFieldClick = { loginViewModel.createAccountDateDialogClicked() },
-                                onDateSubmission = { loginViewModel.createAccountDateDialogSubmitClicked() },
-                                onDateDismiss = { loginViewModel.createAccountDateDialogDismiss() },
+                                onDateFieldClick = { createAccountViewModel.createAccountDateDialogClicked() },
+                                onDateSubmission = { createAccountViewModel.createAccountDateDialogSubmitClicked() },
+                                onDateDismiss = { createAccountViewModel.createAccountDateDialogDismiss() },
                                 onSubmission = {
-                                    loginViewModel.createAccount()
+                                    createAccountViewModel.createAccount()
                                 }
                             ) {
                                 navController.navigateUp()
@@ -84,7 +86,7 @@ class LoginActivity : ComponentActivity() {
                         }
                     }
                 }
-                loginViewModel.onAccountCreateSuccess = {
+                createAccountViewModel.onAccountCreateSuccess = {
                     navController.navigateUp()
                 }
             }
