+++ applegacy/src/main/res/values/attrs.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <attr name="legacy_background" format="color" />
-    <attr name="legacy_primary" format="color" />
-    <attr name="legacy_onPrimary" format="color" />
-    <attr name="legacy_secondaryVariant" format="color" />
-    <attr name="legacy_surface" format="color" />
-    <attr name="legacy_onSurface" format="color" />
-    <attr name="legacy_onError" format="color" />
-    <attr name="legacy_onBackground" format="color" />
+    <attr name="legacyBackground" format="color" />
+    <attr name="legacyPrimary" format="color" />
+    <attr name="legacyOnPrimary" format="color" />
+    <attr name="legacySecondaryVariant" format="color" />
+    <attr name="legacySurface" format="color" />
+    <attr name="legacyOnSurface" format="color" />
+    <attr name="legacyOnError" format="color" />
+    <attr name="legacyOnBackground" format="color" />
 </resources>
