+++ legacyCore/build.gradle
@@ -6,7 +6,7 @@ plugins {
 }
 
 android {
-    namespace 'com.hamdan.forzenbook.createaccountcore'
+    namespace 'com.hamdan.forzenbook.legacy.core'
     compileSdk 33
 
     defaultConfig {
@@ -24,40 +24,23 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_11
-        targetCompatibility JavaVersion.VERSION_11
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
     kotlinOptions {
-        jvmTarget = '11'
+        jvmTarget = '1.8'
     }
 }
 
 dependencies {
-    //Modules
-    implementation project(':core')
-    implementation project(':uiCore')
-
-    // MockK
-    testImplementation("io.mockk:mockk:1.13.4")
-
     // Hilt
     implementation "com.google.dagger:hilt-android:2.44"
     kapt "com.google.dagger:hilt-compiler:2.44"
 
-    // retrofit
-    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-
-    // gson
-    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
-
     implementation 'androidx.core:core-ktx:1.9.0'
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.5.1'
-    implementation 'androidx.activity:activity-ktx:1.6.1'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.google.android.material:material:1.8.0'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-}
-
-kapt {
-    correctErrorTypes true
 }
