+++ legacyCore/src/main/java/com/hamdan/forzenbook/legacy/core/view/Navigator.kt
@@ -1,0 +1,6 @@
+package com.hamdan.forzenbook.legacy.core.view
+
+interface Navigator {
+    fun navigateToLogin(): Unit
+    fun navigateToCreateAccount(): Unit
+}
