+++ loginCore/build.gradle
@@ -36,6 +36,7 @@ dependencies {
 
     //Modules
     implementation project(':core')
+    implementation project(':uiCore')
 
     // MockK
     testImplementation("io.mockk:mockk:1.13.4")
@@ -58,7 +59,7 @@ dependencies {
 
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.5.1'
-    implementation 'androidx.activity:activity-compose:1.7.0-beta01'
+    implementation 'androidx.activity:activity-ktx:1.6.1'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
