+++ settings.gradle
@@ -22,3 +22,4 @@ include ':composeCore'
 include ':createAccountCompose'
 include ':uiCore'
 include ':applegacy'
+include ':legacyCore'
